added comments to syntax module
I'm starting to appreciate more and more the value of commenting code.
I need to scale my coding beyond myself, and the way to do this is not to
do clever tricks that make code hard to read: the way to do this is to
write code that others can understand and build off of.

Ideally, I would be able to write with a proper literate coding tool, like
noweb, but rustdoc will have to do for now because it is so well integrated.