feat: the cli now parses a collection of type definitions
It doesn't do anything with them: it just prints their debug output.

Notes:

- We convert the ParseError thrown from lalrpop into a string
in order to throw it as an error. This is because the ParseError
has a lifetime parameter tied to the source &str.
The default thing that we get here is somewhat ugly, and also could
use annotated source code to be more clear instead of just a raw position.

- We are taking a stance on significant whitespace: namely we are against
it! We also are favoring semicolons over commas, for ease of adding/deleting
without having to add/delete commas, ala nix.

Do next:

- Parse primitive types
- Pretty-print, possibly with [pretty](https://docs.rs/pretty/)
- Kind-check
- Think about content addressing/UUIDs?